feat: Update show_payment_link.html to display payment link logo and tag name
The show_payment_link.html template has been updated to include the display of the payment link logo and tag name. If a payment link logo is available, it will be shown as an image with a width and height of 50px. The tag name will be displayed alongside the logo. This change improves the visual representation of the payment link on the page.