style: Update landing page title and meta tags
The landing page title and meta tags have been updated to improve search engine optimization (SEO) and social media sharing. The title has been changed to "Bixmerchant | Payment With Crypto Made Easy" to better reflect the purpose of the website. Additionally, the meta tags for Facebook and Twitter have been added to provide more information when the page is shared on social media platforms. These changes enhance the visibility and branding of the landing page.